Add cash transaction balance to store
